Rework Makefile 'cps' target to run 'pull_source' command
This set of changes extends the 'make cps' target to run the
'./run.sh pull_source' command for each CP repo it clones.

This change warranted redefining the CP-related Makefile variables
to be more in line with Makefile syntax and target usage.

The use of the hidden ./cp_root/.pulled_%.git target files allows
the Makefile to know if the CP was fully clone and pulled. Otherwise,
the Makefile has no obvious way of detecting which CPs are fully
synced or not.