Use the local ./cp_root directory as a bind mount to the CRS container
This gets the necessary /cp_root volume into the CRS container. It also
stops defining an unused and uninitialized cp_root host-level docker
volume that didn't work or do anything.

This removed the cp_root from the iAPI service, but it is unclear
as to why that was ever needed.