Use a yaml file to configure CP targets for competition
The CPs to initialize in the CP volume are now specified
in a file `cp_config.yaml` that is read by the Makefile
when invoking the `make cps` target.

The yaml file includes the name of CP (which is how the
directory will be named in the CP root directory), the
URL for the git repository, and the optional reference
(tag, branch, or commit) that should be checked out. The
latter will default to "main" if not specified.

The Makefile variable (`CP_CONFIG_FILE`) that specifies
the CP targets yaml file can be overriden with an environment
variable when invoking `make`.